+++ internal/sandbox/mediaimage/embed.go
@@ -1,0 +1,6 @@
+package mediaimage
+
+import _ "embed"
+
+//go:embed Dockerfile
+var dockerfileBytes []byte